refactor(node): extract EngineToPrimaryRpc from lib root
